+++ RELEASES.md
@@ -15,6 +15,7 @@ Version 0.8.17 (2022-XX-XX)
 * Genesis GV70 2022-23 support thanks to zunichky and sunnyhaibin!
 * Hyundai Santa Cruz 2021-22 support thanks to sunnyhaibin!
 * Kia Sportage 2023 support thanks to sunnyhaibin!
+* Kia Sportage Hybrid 2023 support thanks to sunnyhaibin!
 
 Version 0.8.16 (2022-08-26)
 ========================